# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Social Media API
  description: REST API for Social Media Demo Application
  version: 1.0.0
  contact:
    name: Social Media Team
    email: team@socialmedia.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api
    description: Local development server
  - url: https://api.socialmedia.example.com/api
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and return JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: john.doe@example.com
                password:
                  type: string
                  format: password
                  example: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/register:
    post:
      tags:
        - Authentication
      summary: User registration
      description: Register a new user account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
                - displayName
              properties:
                username:
                  type: string
                  example: johndoe
                email:
                  type: string
                  format: email
                  example: john.doe@example.com
                password:
                  type: string
                  format: password
                  example: password123
                displayName:
                  type: string
                  example: John Doe
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Invalid input or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # User endpoints
...
  /posts/{id}:
    get:
      tags:
        - Posts
      summary: Get post by ID
      description: Retrieve a specific post by its ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '200':
          description: Post retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Posts
      summary: Delete a post
      description: Delete a post (only by the author)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '204':
          description: Post deleted successfully
        '403':
          description: Not authorized to delete this post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /posts/{postId}/comments:
    get:
      tags:
        - Comments
      summary: Get post comments
      description: Retrieve comments for a specific post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Comments retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentList'
  /comments:
    post:
      tags:
        - Comments
      summary: Create a comment
      description: Add a comment to a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
                - postId
              properties:
                content:
                  type: string
                  example: "Great post! Thanks for sharing."
                  maxLength: 280
                postId:
                  type: string
                  example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '201':
          description: Comment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /comments/{id}:
    delete:
      tags:
        - Comments
      summary: Delete a comment
      description: Delete a comment (only by the author)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '204':
          description: Comment deleted successfully
        '403':
          description: Not authorized to delete this comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
        username:
          type: string
          example: "johndoe"
        email:
          type: string
          format: email
          example: "john.doe@example.com"
        displayName:
          type: string
          example: "John Doe"
        avatar:
          type: string
          nullable: true
          example: "https://example.com/avatar.jpg"
        bio:
          type: string
          nullable: true
          example: "Software developer passionate about technology"
        followersCount:
          type: integer
          example: 150
        followingCount:
          type: integer
          example: 75
        postsCount:
          type: integer
          example: 42
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    Post:
      type: object
      properties:
        id:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
        content:
          type: string
          example: "Just deployed my first microservice to Kubernetes!"
        authorId:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
        author:
          $ref: '#/components/schemas/User'
        likesCount:
          type: integer
          example: 24
        commentsCount:
          type: integer
          example: 8
        sharesCount:
          type: integer
          example: 3
        isLiked:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    Comment:
      type: object
      properties:
        id:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
        content:
          type: string
          example: "Great post! Thanks for sharing."
        postId:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
        authorId:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
        author:
          $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
...
    CommentList:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 20
        totalElements:
          type: integer
          example: 50
        totalPages:
          type: integer
          example: 3
        first:
          type: boolean
          example: true
        last:
          type: boolean
          example: false
...
    Error:
      type: object
      properties:
        message:
          type: string
          example: "Invalid input provided"
        status:
          type: integer
          example: 400
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        path:
          type: string
          example: "/api/posts"